Separated the search page and 404 page into views
